backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[asyncio]==5.2.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
langgraph==0.2.59
langchain==0.3.13
langchain-google-genai==2.0.7
google-generativeai==0.8.3
langchain-community==0.3.13
chromadb==0.5.23
fhir.resources==7.1.0
httpx==0.28.1
tenacity==9.0.0
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-httpx==0.35.0
faker==33.1.0
python-dotenv==1.0.1
websockets==14.1
anyio==4.7.0
